edit model product - coding add order - fix bugs

+++ lib/widgets/chart_item_widget.dart
@@ -99,7 +99,7 @@ class _ChartItemWidgetState extends State<ChartItemWidget> {
     );
   }
 
-  double getPrice() {
+  int getPrice() {
     return widget.item.price * amount;
   }
 }
